Remove the dep sync15 has on support to avoid it bringing in JNA
