+++ components/sync15/android/build.gradle
@@ -28,7 +28,6 @@ android {
 
 dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation project(':support')
 }
 
 apply from: "$rootDir/publish.gradle"
