+++ components/sync15/android/src/main/java/mozilla/appservices/sync15/SyncTelemetryPing.kt
@@ -4,10 +4,10 @@
 
 package mozilla.appservices.sync15
 
-import mozilla.appservices.support.stringOrNull
-import mozilla.appservices.support.unwrapFromJSON
 import org.json.JSONArray
 import org.json.JSONObject
+import org.json.JSONException
+
 
 /**
  * This file defines Kotlin data classes for unpacking the Sync telemetry ping,
@@ -497,4 +497,25 @@ private fun intOrZero(jsonObject: JSONObject, key: String): Int {
     return unwrapFromJSON(jsonObject) {
         it.getInt(key)
     } ?: 0
+}
+
+/**
+ * Extracts an optional property value from a JSON object, returning `null` if
+ * the property doesn't exist.
+ */
+inline fun <T> unwrapFromJSON(jsonObject: JSONObject, func: (JSONObject) -> T): T? {
+    return try {
+        func(jsonObject)
+    } catch (e: JSONException) {
+        null
+    }
+}
+
+/**
+ * Extracts an optional string value from a JSON object.
+ */
+fun stringOrNull(jsonObject: JSONObject, key: String): String? {
+    return unwrapFromJSON(jsonObject) {
+        it.getString(key)
+    }
 }
